moves joinMaps to Essentials Core
